sentinela/requirements.txt
flask
sqlalchemy
flask-sqlalchemy
flask-login
werkzeug
pytesseract
Pillow
opencv-python
numpy
